[Java] Always update time after a batch has been sent to accommodate for slow senders.
Old code would update time only when `now < timestamp` which would not be true for slow sender thus leading to the missing reporting intervals and late exit from the loop.
For example say we want to send 1000 messages per second and send of message takes 100ms. After the first message the `now` will be updated to `start + 100ms`
whereas `timestamp` will be set to `start + 1ms`. After the second message the `timestamp` will be set to `start + 2ms` but `now` will not be updated, because
the check `now < timestamp` will be false. In this scheme it would take 101 iteration for the check to fire again at which point more than 10 seconds would have elapsed.