[Java] Add support for multiple active channels in echo benchmark.
It is now possible to define multiple destination and source channels (the numbers must match) and the echo benchmark will send messages in a round robin fashion to one of the destination channels at a time and will poll source subscriptions for responses. The total number of sent messages remains the same just the channel used for sending differs for each send attempt.